0.17
Finished both Barracks' although need to add Guard spawns and correct how the stairs are used. Expanded upon cave system benath Newchester, now there are Cave Goblins, Cave Bats, Death Spiders and a few Rats near the enterance. There is a small camp close to the enterance and an open "hallway" further along with an underground river. You can mine stone and flint from the tall rocks; there is also a few coal ore nodes close to the Death Spider spawns. This makes getting the ore risky for low levels and seeing as though it is most likely going to be the first ore I will either replace coal with a higher value ore or move the Death Spiders. Also created two new buildings in the last chunk of the S1 line of Newchester but they need interiors still.