0.22
Finished Cathedral and graveyard with climbable tower. Tower has new ambience as does Cathedral. May add a tomb (whether this enterance is in the Cathedral or the graveyard though is unknown). The entirety of Newchester is now visible, however the new housing area needs interiors. Added a factory with new ambience, added a greenhouse. Added fish (might remove the nametag above them as it is annoying). Added a park and added dogs.