0.23
Extended ocean and forest, added interiors to five houses and created a restaurant with storage shed. Need to add "waiter/waitress" so that players can "order" food. Bug with house 5 means that when a player leaves the building they are teleported out of the map; this will be patched next update.